+++ algorithms/celestialTwoBodyPoint/celestialTwoBodyPointAlgorithm.cpp
@@ -5,10 +5,10 @@
 */
 
 #include "celestialTwoBodyPointAlgorithm.h"
-#include "../freestandingInvalidArgument.h"
 #include "../utilities/safeMath.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "architecture/utilities/rigidBodyKinematics.hpp"
+#include "utilities/freestandingInvalidArgument.h"
 
 void CelestialTwoBodyPointAlgorithm::reset(const bool secCelBodyIsLinked) {
     this->secCelBodyIsLinked = secCelBodyIsLinked;
