+++ algorithms/mrpFeedback/mrpFeedbackAlgorithm.cpp
@@ -2,7 +2,7 @@
 #include "architecture/utilities/eigenSupport.h"
 #include "utilities/timeConstants.h"
 
-#include "../freestandingInvalidArgument.h"
+#include "utilities/freestandingInvalidArgument.h"
 #include <math.h>
 
 /*! This method performs a complete reset of the module.  Local module variables that retain
