+++ algorithms/mrpPD/_tests/mrpPDTestHelpers.hpp
@@ -1,9 +1,9 @@
 #ifndef TEST_MRPPD_H
 #define TEST_MRPPD_H
 
-#include "../freestandingInvalidArgument.h"
 #include "architecture/utilities/eigenSupport.h"
 #include "mrpPDAlgorithm.h"
+#include "utilities/freestandingInvalidArgument.h"
 #include <gtest/gtest.h>
 #include <math.h>
 #include <Eigen/Core>
