+++ algorithms/rateControl/rateControlAlgorithm.cpp
@@ -5,8 +5,8 @@
 */
 
 #include "rateControlAlgorithm.h"
-#include "../freestandingInvalidArgument.h"
 #include "../utilities/validInertiaCheck.h"
+#include "utilities/freestandingInvalidArgument.h"
 
 /*! Computes the required control torque Lr from the attitude rate error and the reference frame angular acceleration.
  @param omega_BR_B the angular velocity of the body frame relative to the reference frame,
